+++ pkg-py/CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -->
 
+## [UNRELEASED]
+
+### Bug fixes
+
+* Fixed a calculation to correctly convert `in` and `cm` to `rem` units for `brand.typography.base.size`. (#60)
+
 ## [0.1.0]
 
 Initial release of `brand_yml`.